REF_CODE_REFACTOR: Reshape refactring the reference code for reshape operator.
BUG=refactoring existing code.
