+++ tensorflow/lite/micro/kernels/BUILD
@@ -166,6 +166,7 @@ cc_library(
         "quantize_common.cc",
         "reduce.cc",
         "reshape.cc",
+        "reshape_common.cc",
         "resize_bilinear.cc",
         "resize_nearest_neighbor.cc",
         "round.cc",
@@ -199,6 +200,7 @@ cc_library(
         "micro_ops.h",
         "pooling.h",
         "quantize.h",
+        "reshape.h",
         "softmax.h",
         "svdf.h",
     ],
