+++ tensorflow/lite/micro/kernels/reshape_test.cc
@@ -39,7 +39,7 @@ void ValidateReshapeGoldens(TfLiteTensor* tensors, int tensors_size,
                             int* expected_dims, const size_t expected_dims_len,
                             bool expect_failure) {
   const TfLiteRegistration registration =
-      tflite::ops::micro::Register_RESHAPE();
+      tflite::Register_RESHAPE();
   micro::KernelRunner runner(registration, tensors, tensors_size, inputs_array,
                              outputs_array,
                              /*builtin_data=*/nullptr);
